requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
fake-useragent==1.4.0
cloudscraper==1.2.71
html5lib==1.1
gunicorn==21.2.0
flask-cors==4.0.0
PyJWT==2.8.0
